requirements.txt
asn1crypto==0.24.0
astroid==2.0
autopep8==1.4.1
bcrypt==3.1.5
captcha==0.2.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
ddt==1.2.0
diff-match-patch==20121119
Django==2.0.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-filters==0.2.1
django-formtools==2.1
django-import-export==1.0.1
django-reversion==2.0.13
djangorestframework==3.9.0
elasticsearch==6.3.1
et-xmlfile==1.0.1
future==0.16.0
geographiclib==1.49
geopy==1.18.1
get==1.0.3
gitdb2==2.0.5
GitPython==2.1.11
httplib2==0.9.2
idna==2.7
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.4
paramiko==2.4.2
Pillow==5.2.0
post==1.0.2
public==1.0.3
pyasn1==0.4.5
pycodestyle==2.4.0
pycparser==2.19
pylint==2.0.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pytz==2018.5
PyYAML==4.2b4
query-string==1.0.2
request==1.0.2
requests==2.19.1
robotframework==3.0.4
robotframework-requests==0.5.0
schedule==0.5.0
six==1.11.0
smmap2==2.0.5
sshtunnel==0.1.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
